Added another missed change
